+++ Changelog.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.8.0] - TBD
+### Added
+* ApiPlatformTestCase now has constants for RFC 7807 problem responses
+* ApiPlatformTestCase::testOperation now has a 'postFormAuth' option
+* `FormDecoder` to decode 'application/x-www-form-urlencoded' requests
+
 ## [2.7.0] - 2023-09-11
 ### Added
 * `ArrayUtil` class with `mergeValues` to merge 2 or more arrays and keep